requirements.txt
Flask>=3.0.0
spotipy>=2.23.0
requests>=2.31.0
python-dotenv>=1.0.0
youtube-search-python>=1.6.6